refactor(Multiselect): DRY (renderListItem)
